move functions for post processing and pdf report creation to new files
